+++ src/optree.cpp
@@ -93,28 +93,28 @@ void BuildModule(py::module& mod) {  // NOLINT[runtime/references]
 
     PyTreeSpecTypeObject
         .def_property_readonly(
-            "num_leaves", &PyTreeSpec::num_leaves, "Number of leaves in the tree.")
+            "num_leaves", &PyTreeSpec::GetNumLeaves, "Number of leaves in the tree.")
         .def_property_readonly("num_nodes",
-                               &PyTreeSpec::num_nodes,
+                               &PyTreeSpec::GetNumNodes,
                                "Number of nodes in the tree. "
                                "Note that a leaf is also a node but has no children.")
         .def_property_readonly("num_children",
-                               &PyTreeSpec::num_children,
+                               &PyTreeSpec::GetNumChildren,
                                "Number of children in the current node. "
                                "Note that a leaf is also a node but has no children.")
         .def_property_readonly(
             "none_is_leaf",
-            &PyTreeSpec::get_none_is_leaf,
+            &PyTreeSpec::GetNoneIsLeaf,
             "Whether to treat None as a leaf. "
             "If false, None is a non-leaf node with arity 0. "
             "Thus None is contained in the treespec rather than in the leaves list.")
         .def_property_readonly(
             "namespace",
-            &PyTreeSpec::get_namespace,
+            &PyTreeSpec::GetNamespace,
             "The registry namespace used to resolve the custom pytree node types.")
         .def_property_readonly(
             "type",
-            &PyTreeSpec::get_type,
+            &PyTreeSpec::GetType,
             "The type of the current node. Return None if the current node is a leaf.")
         .def("unflatten",
              static_cast<py::object (PyTreeSpec::*)(const py::iterable&) const>(
@@ -123,13 +123,13 @@ void BuildModule(py::module& mod) {  // NOLINT[runtime/references]
              py::arg("leaves"))
         .def("flatten_up_to",
              &PyTreeSpec::FlattenUpTo,
-             py::arg("full_tree"),
              "Flatten the subtrees in ``full_tree`` up to the structure of this treespec "
-             "and return a list of subtrees.")
+             "and return a list of subtrees.",
+             py::arg("full_tree"))
         .def("compose",
              &PyTreeSpec::Compose,
-             py::arg("inner_treespec"),
-             "Compose two treespecs. Constructs the inner treespec as a subtree at each leaf node.")
+             "Compose two treespecs. Constructs the inner treespec as a subtree at each leaf node.",
+             py::arg("inner_treespec"))
         .def("walk",
              &PyTreeSpec::Walk,
              "Walk over the pytree structure, calling ``f_node(children, node_data)`` at nodes, "
@@ -137,9 +137,19 @@ void BuildModule(py::module& mod) {  // NOLINT[runtime/references]
              py::arg("f_node"),
              py::arg("f_leaf"),
              py::arg("leaves"))
+        .def("is_prefix",
+             &PyTreeSpec::IsPrefix,
+             "Test whether this treespec is a prefix of the given treespec.",
+             py::arg("other"),
+             py::arg("strict") = true)
+        .def("is_suffix",
+             &PyTreeSpec::IsSuffix,
+             "Test whether this treespec is a suffix of the given treespec.",
+             py::arg("other"),
+             py::arg("strict") = true)
         .def("children", &PyTreeSpec::Children, "Return a list of treespecs for the children.")
         .def("is_leaf",
-             &PyTreeSpec::is_leaf,
+             &PyTreeSpec::IsLeaf,
              "Test whether the current node is a leaf.",
              py::arg("strict") = true)
         .def("__str__", &PyTreeSpec::ToString, "Return a string representation of the treespec.")
@@ -157,10 +167,34 @@ void BuildModule(py::module& mod) {  // NOLINT[runtime/references]
             py::is_operator(),
             py::arg("other"))
         .def(
+            "__lt__",
+            [](const PyTreeSpec& a, const PyTreeSpec& b) { return a < b; },
+            "Test for this treespec is a strict prefix of another object.",
+            py::is_operator(),
+            py::arg("other"))
+        .def(
+            "__le__",
+            [](const PyTreeSpec& a, const PyTreeSpec& b) { return a <= b; },
+            "Test for this treespec is a prefix of another object.",
+            py::is_operator(),
+            py::arg("other"))
+        .def(
+            "__gt__",
+            [](const PyTreeSpec& a, const PyTreeSpec& b) { return a > b; },
+            "Test for this treespec is a strict suffix of another object.",
+            py::is_operator(),
+            py::arg("other"))
+        .def(
+            "__ge__",
+            [](const PyTreeSpec& a, const PyTreeSpec& b) { return a >= b; },
+            "Test for this treespec is a suffix of another object.",
+            py::is_operator(),
+            py::arg("other"))
+        .def(
             "__hash__",
             [](const PyTreeSpec& t) { return absl::HashOf(t); },
             "Return the hash of the treespec.")
-        .def("__len__", &PyTreeSpec::num_leaves, "Number of leaves in the tree.")
+        .def("__len__", &PyTreeSpec::GetNumLeaves, "Number of leaves in the tree.")
         .def(py::pickle([](const PyTreeSpec& t) { return t.ToPicklable(); },
                         [](const py::object& o) { return PyTreeSpec::FromPicklable(o); }),
              "Serialization support for PyTreeSpec.");